prevent gvim focus click at eclipse startup
If a gvim instance was left open at the last eclipse shutdown, then the
next startup will open it and attempt to focus. But the focus call
occurs before the shell has been displayed, leading to 0,0 bounds result
and a click at that location on the user's screen, regardless of where
eclipse will render.

git-svn-id: https://eclim.svn.sourceforge.net/svnroot/eclim/trunk@2238 b5bd0a2e-1a23-0410-8567-9814cbbba304