improve local eclipse detection
on Fedora the eclipse local dir is something like ~/.eclipse/1156789,
which the current detection code would not pick up, so change it to look
for any directory in .eclipse that looks like an eclipse path (has a
configuration dir in it).  Also, log should we find more than one
candidate or raise an error dialog if after initializing eclipse, we
find none.