make org.eclim.Eclim.resolveFile() more robust
Often projects are available from a workspace, but their source code
is not in a sub-directory of the workspace directory: this happens
after an import "Existing projects into Workspace" and "copy projects
into workspace" is not selected. This commit allows Eclim to detect
this and return a correct result in all cases.