Adds usage() that prints all possible commands with their arguments.
The commit is a bit big as it also adds a unit test checking the
result of usage() which itself requires  plumbing to get all
commands. I hope this plumbing will get reused in more unit tests
which makes it worth it. Moreover, a dependency to the reflections
library is added which requires guava and javassist. Again, this
dependency might be worth it as they could replace the scannotation
library currently used (which is both unmaintained and harder to use
according to the author of scannotation).