on complete write the temp file when getting starting index
vim modifies the file between the call to get the start index and the
call to get the results, so saving on the call to get the results breaks
completing based on the current prefix.

also disable temp file for php completion since pdt seems to not handle
the temp file well.