remove jdt formatter settings now that everything works by default
since all eclipse code changes now go through the formatter and a retab
on the vim side, the previous tab vs spaces issues between vim and
eclipse are no longer a problem.

also, recent versions of vim recognize edit as a change in the undo
tree, so all our "gotachas" are now resolved.