add tab ids set on buffers for prevent close + :Buffers
set a unique id on each tab and store that id for every buffer in each
tab so that the "prevent clone on buffer delete" can open a relevant
buffer previously opened on that tab instead of some other buffer no
longer open from work on some other tab.

also allows :Buffers to limit its results to the current tab.