update c/cpp validation to use CodanRunner for non-syntactic problems
Changes in the gui editor ultimately use the CodanRunner so lets use it
from eclim as well. Something I didn't notice before is that the old
semantic error processing resulted in ultimately the same errors as the
editor, but with different error messages. After switching to
CodanRunner, the error messages now match up.

Should resolve #156 and the related pull request, #158