when reading an archived file, preserve vim's alternate file
when opening an archived file, the file is first set to the archive url,
then to the resulting temp file path. This prevents <c-^> from jumping
back to the previous file since in this case the previous file is the
archived file url. Use 'keepalt' when switching from the archived url to
the temp file so that <c-^> and other alternate file features work as
expected.