update adt build script to only extract + compile against greatest adt version found
when upgrading plugins, eclipse will keep the old jars around, so eclim
needs to be able to compile against the correct ones. we'll just assume
that highest version number is the correct one since invoking eclipse
(which is slow) is the only way to know for sure which plugin version is
active.

relates to #176