change format of news dates to fix rss+sphinx
If the section title, with non-alphanumrics removed, is all numbers,
then sphinx generates its own incrementing section ids for the page.
This is a problem for the rss extension since the same url is reused for
a different news items as items are added/removed, which can confuse rss
readers attempting to determine if there are new entries in the feed.