fix :ProjectTree use of fugitive for diplaying branch info
the ProjectTree has a User autocmd to trigger the tree's vcs info, which
will then trigger a fugitive autocmd, which will trigger a User autocmd,
which then inadvertently triggers the ProjectTree User autocmd, and so
on until vim says enough is enough. Fix this by preventing the
ProjectTree autocmd from being executed recursively.