fix eclimd script to only set jvm arch if native swt jar is found
on linux distros like ubuntu (and probably debian), the swt native
libraries are a separate package install outside of the eclipse plugins
location. The eclimd script will not find those jars which led the
current logic to fallback into -d32, which could be incorrect and
prevent eclimd from starting on 64bit machines.