Implemented CallHierarchy for Java (caller/callee)
Included eclim command, vim scripts, vim documentation and some attempt
at testing.

I was unable to run the tests as no matter what target I used for ant,
it constantly tried to overwrite files in my eclipse folder. Will try
to look at it again with more attention tomorrow. In any case, I tested
manually around the eclim source code and it appears to be working.

I based the code on C's CallHierarchy. I left the maximum depth at 3
since that's what appeared on the C example but I think it would make
more sense to set it at 2 because with big projects it takes quite some
time. We could then employ a system on the vimscript to ask for more
call hierarchy on a particular element like what is done in the GUI. I'm
not very experienced with VimScript but it seems doable (code similar to
NerdTree, etc).