only place loclist signs when eclim has set the loclist
When the user has syntastic, or some other sign placing validator, for a
filetype that eclim isn't validating/updating, then as the user switches
windows, eclim will still trigger its sign updates on the location list
it is not managing, so if errors are fixed, the eclim signs won't be
removed until the user switches windows again to trigger eclim's signs
update.

This change will prevent eclim from placing signs for a buffer's
location list that eclim hasn't managed at least once (we shouldn't need
to validate that the current list is one we've set, but if a use case
arises that requires that, we can use the same mechanism as the
:ProjectProblems command uses to validate ownership of the quickfix
list).

refs #241