fix installer vimfiles feature set to handle new versioned feature support
when adding support for versioned features, specifically for sdt, I
neglected to test that the vimfiles were installed properly, which they
weren't due to a pattern set for **/org.eclim.sdt10 being added instead
of **/org.eclim.sdt. Updating the code to add both to hopefully handle
both versioned features and perhaps the possibility of a non-versioned
feature with a number suffix.

fixes #243