RubyProjectManager.refresh(IProject, IFile): ignore IllegalArgumentExceptions
if the project is mixing more than one nature and a ProjectUtils.getFile
call is made on say a js file outside of any of the ruby src dirs, then
the subsiquent call to the ruby project manager would raise an
exception. For the sake of refresh, we'll ignore those exceptions.

refs #244