Add -a <action> parameter to JavaSearch and JavaSearchContext
The parameter allows to specify where the search result should be
opened, e.g. vsplit would open the result in a vertical split. If the
action parameter is not specified it defaults to g:EclimJavaSearchSingleResult,
which itself defaults to g:EclimDefaultFileOpenAction.

This parameter works for the JavaSearch and JavaSearchContext commands
and in the case of JavaSearchContext it is the only parameter.

Examples:
Open single search result for 'Foo' in a vertical split:
  JavaSearch -p Foo -a vsplit
Open single search result for 'Foo' in the current window:
  JavaSearch -p Foo -a edit