update eclim created signs to have non-overlapping ids by sign name
this allows more than one sign of different names to be placed on a
line. while only the last sign's text will be displayed, removing of
that sign will not remove all other signs on that line, so the next
available sign on the line will be still be displayed.