remove choose prompt from multiple search response
when moving the handling of search responses to lang, the behavior of
multiple search results changed to prompting the user to choose a
result. After going through the code history, I could find no clue as to
why this bit of code ever existed (previously only trigger when using
the -a arg to search).