fix JavaUtils.getPackageFragment to match full package name
If you have an existing package "foo.bar" and used getPackageFragment
with "bar", it would return "foo.bar" as a found package since it is
just checking if the path on disk ends in the supplied package name.
Instead, check the actual package name and don't use the on disk path.

refs #362