Refactoring and initial async execution work
If async execution is not possible, it tries to fallback on
synchronous execution. However, it will now terminate the launch
and return an error if that execution spawned long-lived processes.
For example, running an Android project does not leave any processes
behind, so it can be run synchronously even if the vim used doesn't
have remote-send support.

Still TODO: tracking the launches, opening the output buffer, handling
process death (IE: it dies on its own) and terminating processes.
Perhaps I'm missing something, but I can't seem to find a way to hook
into the IProcess's life cycle....