JavaSearchContext command should look for references.
As per the documentation, only references should be returned when the
cursor is on method/field declaration. Also, this makes sense from
usability perspective as we don't need to include the current line in
the results.

Testing:
Verified that current line is not included in result set and only
references are returned when cursor in on a public method.