fix :ProjectRun output window re-use
previous behavior was to rename the output window when the program
terminated, then on subsequent run, look for that window, delete the
buffer then create a new output window.

this caused problems, at least using console vim on linux, where the
buffer deletion would not be immediate since it occured from a remote
command call. this resulted in the server side failing to capture the
output buffer number and raising an exception. There may very well have
been some error other error on the vim side not being exposed, but it
seemed to be easier to just re-use the window. If the program status
needs to be re-added later, that can go into the statusline instead to
avoid buffer name issues.

refs #354