GroovyCompleteJavaDocURI
We changed the 'groovy_complete' command:
The client can now request java doc links for each of the suggestions.

How:
We added a new flag 'javaDoc' to the 'groovy_complete' command. If this flag is set to 'true', the URI link which points to the java doc of the corresponding suggestion is returned in an additional field 'javaDocURI' in each CodeCompleteResult. If no URI can be created, the 'javaDocURI' field is set to an empty string.
If the flag is set to 'false', the 'javaDocURI' is set to an empty string.